[move] Service登録ページの移動
